refactor: move unified spec runner to tools and create integration functional test placeholder directory
