+++ package.json
@@ -103,7 +103,7 @@
     "check:eslint": "eslint -v && eslint --max-warnings=0 --ext '.js,.ts' src test",
     "check:tsd": "tsd --version && tsd",
     "check:dts": "./node_modules/typescript/bin/tsc --noEmit mongodb.d.ts && tsd",
-    "check:test": "mocha --file test/tools/runner --recursive test/functional",
+    "check:test": "mocha --file test/tools/runner --recursive test/functional test/integration",
     "check:unit": "mocha --recursive test/unit/",
     "check:ts": "./node_modules/typescript/bin/tsc -v && ./node_modules/typescript/bin/tsc --noEmit",
     "check:atlas": "mocha --config \"test/manual/mocharc.json\" test/manual/atlas_connectivity.test.js",
