+++ test/functional/change_stream_spec.test.js
@@ -3,7 +3,7 @@
 const path = require('path');
 const chai = require('chai');
 const { loadSpecTests } = require('../spec');
-const { runUnifiedSuite } = require('./unified-spec-runner/runner');
+const { runUnifiedSuite } = require('../tools/unified-spec-runner/runner');
 const camelCase = require('lodash.camelcase');
 const { setupDatabase } = require('./shared');
 const { delay } = require('./shared');
