+++ test/functional/collection_management_spec.test.js
@@ -2,7 +2,7 @@
 
 const { expect } = require('chai');
 const { loadSpecTests } = require('../spec/index');
-const { runUnifiedTest } = require('./unified-spec-runner/runner');
+const { runUnifiedTest } = require('../tools/unified-spec-runner/runner');
 
 describe('Collection management unified spec tests', function () {
   for (const collectionManagementTest of loadSpecTests('collection-management')) {
