+++ test/functional/crud_spec.test.js
@@ -7,7 +7,7 @@ const expect = chai.expect;
 chai.use(require('chai-subset'));
 
 const { loadSpecTests } = require('../spec/index');
-const { runUnifiedTest } = require('./unified-spec-runner/runner');
+const { runUnifiedTest } = require('../tools/unified-spec-runner/runner');
 
 function enforceServerVersionLimits(requires, scenario) {
   const versionLimits = [];
